Implement cloud image storage functionality
The changes were made to implement cloud image storage functionality
in the application. This allows users to upload and store their profile
pictures in the cloud, providing a more scalable and reliable
solution compared to storing the images locally on the server.

- The `UserIconMenu.jsx` component was modified to add functionality
related
to user icons in the navigation bar. This allows users to easily access
their profile and perform actions related to their account.

- The `ProfilePage.jsx` component was modified to implement new features
and fix bugs in the user profile page. This includes adding the
ability for users to upload and display their profile picture from the
cloud.

- The `auth-controller.js` file was modified to add authentication logic
for cloud image storage. This ensures that only authenticated users can
upload and access their profile pictures.

- The `fileHashing.js` file was deleted as it was no longer needed f
or cloud image storage.

- The `googleCloudStorage.js` file was added to handle interactions
with Google Cloud Storage. This includes functions for uploading,
retrieving, and deleting profile pictures.

- The `multer.js` file was modified to configure file upload settings
for cloud image storage. This includes setting the destination and
file naming conventions for uploaded profile pictures.

- The `server.js` file was modified to implement cloud image storage
functionality. This includes handling API endpoints for uploading,
retrieving, and deleting profile pictures.

- The `user-controller.js` file was modified to add functionality
related to user profile pictures. This includes adding the ability for
users to upload and display their profile picture from the cloud.