Fix Publish Bug & drafts, published maps bugs
Fixed the bug where i was checking length when drafts & published maps
were null

Fixed Bug trying to publish right after saving since we are only
returning id's fields were null in the slice. publish then
published empty data. I tried returning the whole object for each
save but run into issues, so instead i am rendering the publish
button when the slice is updated by checking thumbnail url is in the
slice. This is a hacky fix but it works for now.
Tested much as i can everything seems to work. so when user clicks
on their draft we can publish, or when they save again which is updating
and after update we do have our slice updated with the thumbnail url
and all the other necessary fields to publish.
```